requirements.txt
aiokafka==0.8.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
black==23.3.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
kafka-python==2.0.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
